main: add CPU options to larger machines
Some machines are not available at every machine size. Specifically, the
i3en machine type is not available at the 4xlarge size which we use for
every other test. In order for the number of CPUs to be comparable then
we need to be able to specify the number of CPUs and threads per CPU on
these larger machines. This is what this commit does.